+++ configs/testing/scattermoe.yml
@@ -44,8 +44,8 @@ model_args:
     add_bias: false
     position_embedding_type: learned_absolute
     rope_theta: 10000
-    use_scattermoe: true
   attention_implementation: sdpa # flash_attention_2
+  moe_implementation: scattermoe
   # use_padding_free_transformer: true
 
 tuning_args:
