+++ dolomite_engine/hf_models/models/moe_dolomite/config.py
@@ -37,7 +37,6 @@ class MoEDolomiteConfig(CommonConfig):
         m_residual: float | None = None,
         num_experts: int = 8,
         num_experts_per_tok: int = 2,
-        use_scattermoe: bool = False,
         normalize_expert_weights: bool = True,
         output_router_logits: bool = False,
         router_aux_loss_coef: float = 0.001,
@@ -78,7 +77,6 @@ class MoEDolomiteConfig(CommonConfig):
 
         self.num_experts = num_experts
         self.num_experts_per_tok = num_experts_per_tok
-        self.use_scattermoe = use_scattermoe
         self.normalize_expert_weights = normalize_expert_weights
 
         self.output_router_logits = output_router_logits
